feat: add recheck

+++ Mine3AuthProxy/src/main/java/com/boomchanotai/mine3AuthProxy/listeners/PlayerDisconnectListener.java
@@ -18,6 +18,10 @@ public class PlayerDisconnectListener implements AwaitingEventExecutor<Disconnec
 
         return EventTask.async(() -> {
             Address address = RedisRepository.getAddress(event.getPlayer().getUniqueId());
+            if (address == null) 
+                return;
+            
+
             RedisRepository.removePlayer(address);
         });
     }
